Serializer: Fixed ApplicationPacketSerializer to serialize/deserialize properly.

+++ src/inet/common/ResultFilters.cc
@@ -51,7 +51,7 @@ Register_ResultFilter("appPkSeqNo", ApplicationPacketSequenceNumberFilter);
 void ApplicationPacketSequenceNumberFilter::receiveSignal(cResultFilter *prev, simtime_t_cref t, cObject *object, cObject *details)
 {
     if (auto msg = dynamic_cast<ApplicationPacket*>(object))
-        fire(this, t, msg->getSequenceNumber(), details);
+        fire(this, t, (long)msg->getSequenceNumber(), details);
 }
